+++ src/components/Layout.tsx
@@ -2,6 +2,7 @@ import { Component, Show, onMount, onCleanup, createSignal } from 'solid-js'
 import { listen } from '@tauri-apps/api/event'
 import { open as openDialog } from '@tauri-apps/plugin-dialog'
 import { Sidebar } from './Sidebar'
+import { UpdateBanner } from './UpdateBanner'
 import { TaskPanel } from './TaskPanel'
 import { SettingsPage, selectSettingsSection, setSettingsSaveRequested } from './SettingsPage'
 import { ArchivedPage } from './ArchivedPage'
@@ -241,15 +242,18 @@ export const Layout: Component = () => {
         />
       </Show>
 
-      <Show when={showSettings()}>
-        <SettingsPage />
-      </Show>
-      <Show when={showArchived() && !showSettings()}>
-        <ArchivedPage />
-      </Show>
-      <Show when={!showSettings() && !showArchived()}>
-        <TaskPanel />
-      </Show>
+      <div class="flex flex-col flex-1 min-w-0 overflow-hidden">
+        <UpdateBanner />
+        <Show when={showSettings()}>
+          <SettingsPage />
+        </Show>
+        <Show when={showArchived() && !showSettings()}>
+          <ArchivedPage />
+        </Show>
+        <Show when={!showSettings() && !showArchived()}>
+          <TaskPanel />
+        </Show>
+      </div>
 
       <NewTaskDialog
         open={!!newTaskProjectId()}